Initial codebase upload

+++ src/crypto/sha256.h
@@ -21,11 +21,11 @@ public:
     	FinalizeNoPadding(hash, true);
     };
     CSHA256& Reset();
-
+    
 private:
     uint32_t s[8];
     unsigned char buf[64];
-    size_t bytes;
+    uint64_t bytes;
     void FinalizeNoPadding(unsigned char hash[OUTPUT_SIZE], bool enforce_compression);
 };
